# packages/web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Royal Battalion API
  description: |
    REST API for the Royal Battalion gaming community management platform.
    Handles user authentication (Discord OAuth), whitelist management,
    ticket/prospect tracking, match history, server administration,
    and Discord bot integration.

    ## Versioning
    All REST endpoints are versioned and served under the `/v1` prefix, which is
    baked into the `servers[].url` base below (so the paths documented here are
    relative to `/v1`, e.g. `/users` resolves to `/v1/users`).

    A few root-level, NON-versioned endpoints are exempt (external / infra
    contracts that must not move) and declare their own root `servers` override:
    - `GET /health` -- infra/Railway health probe.
    - `GET /admins.cfg` -- fetched by the Squad game server at a fixed URL.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via the `/auth/sync` endpoint.
    The flow is:
    1. User authenticates with Discord OAuth on the frontend
    2. Frontend sends the Discord access token to `POST /auth/sync`
    3. API returns a JWT (valid 4 hours) containing userId and permissions
    4. Include the JWT in all subsequent requests as `Authorization: Bearer <token>`

    ## Permissions
    Access is controlled by a granular RBAC system with 19 permissions.
    Users with the `admin` permission bypass all permission checks.

    ## Rate Limiting
    - Global: 200 requests/minute per IP
...

...
  version: 1.0.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /clans:
    get:
      tags: [Clans]
      summary: List clans
      responses:
        "200":
          description: Clans
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Clan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
    post:
      tags: [Clans]
      summary: Create clan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, tag]
              properties:
                name:
                  type: string
                  minLength: 1
                tag:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Clan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Clan"
        "400":
          description: Clan name or tag already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /clans/{id}:
    patch:
      tags: [Clans]
      summary: Update clan
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                tag:
                  type: string
      responses:
        "200":
          description: Updated clan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Clan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Clan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist
    delete:
      tags: [Clans]
      summary: Delete clan
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Clan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist

  # ── Tickets ───────────────────────────────────────────
...
  /discord-bot/messages:
    get:
      tags: [Discord Bot]
      summary: Query bot messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author Discord ID
        - name: channel
          in: query
          schema:
            type: string
          description: Filter by channel ID
        - name: dm
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Filter to DMs only
        - name: search
          in: query
          schema:
            type: string
          description: Search message content (LIKE match)
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Filter messages after this date
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Filter messages before this date
      responses:
        "200":
          description: Paginated messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PaginatedBotMessages"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - view:discord-bot
        - manage:discord-bot
  /discord-bot/logs:
    get:
      tags: [Discord Bot]
      summary: Query bot logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: level
          in: query
          schema:
            type: integer
          description: Filter by log level
        - name: module
          in: query
          schema:
            type: string
          description: Filter by module name
        - name: search
          in: query
          schema:
            type: string
          description: Search log messages (LIKE match)
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PaginatedBotLogs"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - view:discord-bot
        - manage:discord-bot
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/sync
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Success (no content)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    Paginated:
      type: object
      description: Standard pagination envelope for growing collections.
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean
...
    ApiResponse_PaginatedBotMessages:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: "#/components/schemas/Paginated"
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/BotMessage"
    ApiResponse_PaginatedBotLogs:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          allOf:
            - $ref: "#/components/schemas/Paginated"
            - type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/BotLog"
...
    BotMessage:
      type: object
      properties:
        id:
          type: integer
        messageId:
          type: string
        channelId:
          type: string
        channelName:
          type: string
          nullable: true
        guildId:
          type: string
          nullable: true
        authorId:
          type: string
        authorTag:
          type: string
        content:
          type: string
          nullable: true
        attachments:
          type: array
          items: {}
          nullable: true
        isDm:
          type: boolean
        direction:
          type: string
          enum: [incoming, outgoing]
        createdAt:
          type: string
          format: date-time
    BotLog:
      type: object
      properties:
        id:
          type: integer
        level:
          type: integer
        levelLabel:
          type: string
        module:
          type: string
          nullable: true
        message:
          type: string
        data:
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Clan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tag:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Member comments / profile ───────────────────────
...